pyproject.toml
[tool.poetry]
name = "githubclassificationdataset"
version = "0.1.0"
description = ""
authors = ["Cezar Sas <cezar.sas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
hydra-core = "^1.1.1"
pandas = "^1.3.4"
gensim = "^4.1.2"
scikit-learn = "^1.0.1"
tqdm = "^4.62.3"
nltk = "^3.6.5"
scikit-multilearn = "^0.2.0"
ratelimit = "^2.2.1"
PyGithub = "^1.55"
matplotlib = "^3.4.3"
wikipedia = "^1.4.0"
reconciler = "^0.2.1"
hydra-joblib-launcher = "^1.1.5"
dedupe = "^2.0.8"
Unidecode = "^1.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
